# Add aggregate options